==> Supportive Material/4-Plots learningcurves lot and lit LL.R <==
# Authors: Job Schepens & Frans van der Slik
# Script for: 
# van der Slik, F., Schepens, J., Bongaerts, T, .van Hout, R. 
# Revisiting Hartshorne, Tenenbaum, and Pinker (2018): 
# Critical Period Claim Revisited: Re-analysis of Hartshorne, Tenenbaum, and Pinker (2018) 
# Suggests Steady Decline and Learner-Type Differences.
# Language learning 

# Original script based on: 
# Hartshorne, J. K., Tenenbaum, J. B., & Pinker, S. (2018). 
# A critical period for second language acquisition: Evidence from 2/3 million English speakers. 
# Cognition, 177, 263-277. https://doi.org/10.1016/j.cognition.2018.04.007
#
# Generates plots based on discontinuous models for immersion (lot) and nonimmersion learners (little)
#

library(ggplot2)
library("grid")

PARlot <- c(15.59, .15,  .08,  8.92)  # imm E lot = 1.0
PARlit <- c(18.63, .06,  .10,  -3.34)  # nonimm E little = 1.0
E <- c(1, 1) #  immersion and non immersion

logodds = T
load("binned_old.Rda")  # Generated by "1A-Analyses All groups continu and discontinu LL.R"

binned <- binned_old[binned_old$Eng_little == "lot" | binned_old$Eng_little == "little", ]
toplot <- binned
toplot$experience <- toplot$e
toplot$te <- as.integer(as.character(toplot$te))
type <- "exp.flat"
colnames(toplot)[4]<-"score"
toplot$E <- 1
toplot$E[toplot$Eng_little=="little"]<-E[2]
toplot$E[toplot$Eng_little=="lot"]<-E[1]

#if (type=="exp.flat"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.flat(toplot$age[i],toplot$te[i],toplot$E[i],PAR)}}
#if (type=="exp.disc"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.disc(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3])}}

lottoplot <- toplot[toplot$Eng_little=="lot",]
littoplot <- toplot[toplot$Eng_little=="little",]
if (type=="exp.cont"){
	for (i in 1:length(lottoplot$score)){
	lottoplot$score[i]<-exp.cont(
		lottoplot$age[i],
		lottoplot$te[i],
		lottoplot$E[i],
		PARlot[1],PARlot[2],PARlot[3],PARlot[4])}}
if (type=="exp.cont"){
	for (i in 1:length(littoplot$score)){
	littoplot$score[i]<-exp.cont(
		littoplot$age[i],
		littoplot$te[i],
		littoplot$E[i],
		PARlit[1],PARlit[2],PARlit[3],PARlit[4])}}
	
if (type=="exp.disc.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.disc.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3])}}
if (type=="exp.cont.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.cont.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
if (type=="pow.flat"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.flat(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2])}}
if (type=="pow.disc"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.disc(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
if (type=="pow.cont"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.cont(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
if (type=="pow.disc.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.disc.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3])}}
if (type=="pow.cont.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.cont.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
	
toplot <- rbind(littoplot, lottoplot)	
toplot$te<-as.factor(toplot$te)
toplot$fac<-as.factor(paste(toplot$Eng_little,toplot$te,sep=""))
	
type <- "exp.flat"
p<-ggplot(data=toplot[toplot$Eng_little=="lot",],aes(x=age,y=score))+geom_line(aes(color=factor(fac)),alpha=1)+scale_color_manual(values=c("#009E73","#009E73","#009E73","#D55E00","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#D55E00","#D55E00","#D55E00"))
for (i in seq(2,29,3)){
	 jitter<-0
	 if (type=="exp.flat"){
		 numage<-min(toplot$age[toplot$Eng_little=="lot" & toplot$te==i])
	 }
	 p<-p+annotate("text",x=numage,y=toplot$score[toplot$Eng_little=="lot" & toplot$te==i & (toplot$age==numage)],label=as.character(i),color="black",size=3.5)
}
p<-p+guides(color=FALSE)+labs(y="accuracy (log-odds)",x="age")+theme(plot.title=element_text(size=11,color="black"),axis.title=element_text(size=14,color="black"),axis.text=element_text(size=11,color="black"))
if(logodds){
	p<-p+scale_y_continuous(lim=c(1,4))
}else{
	p<-p+scale_y_continuous(lim=c(.75,1))
}
p<-p+scale_x_continuous(lim=c(0,71)) +  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"))
p	

p2<-ggplot(data=toplot[toplot$Eng_little=="little",],aes(x=age,y=score))+geom_line(aes(color=factor(fac)),alpha=1)+scale_color_manual(values=c("#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#D55E00","#D55E00","#D55E00","#D55E00","#D55E00","#D55E00"))
for (i in seq(4,30,1)){
  jitter<-0
  if (i==5 | i==8){jitter<-11}
  if (i==9){jitter<-16}
  if (i==11){jitter<-5}
  if (i==15){jitter<-4}
  if (i==22){jitter<-14}
  if (i==24){jitter<-5}
  if (i==26){jitter<-11}
  if (i==29){jitter<-3}
  p2<-p2+annotate("text",x=max(toplot$age[toplot$Eng_little=="little" & toplot$te==i])-jitter,y=toplot$score[toplot$Eng_little=="little" & toplot$te==i & (toplot$age==max(toplot$age[toplot$Eng_little=="little" & toplot$te==i])-jitter)],label=as.character(i),color="black",size=3.5)
}
p2<-p2+guides(color=FALSE)+labs(y="accuracy (log-odds)",x="age")+theme(plot.title=element_text(size=11,color="black"),axis.title=element_text(size=14,color="black"),axis.text=element_text(size=11,color="black"))
if(logodds){
	p2<-p2+scale_y_continuous(lim=c(1,4))
}else{
	p2<-p2+scale_y_continuous(lim=c(.75,1))
}
p2<-p2+scale_x_continuous(lim=c(0,71)) +  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"))
p2

source("2-aggregate_functions LL.R")
load("binned_old.Rda")

binned <- binned_old[binned_old$Eng_little == "lot" | binned_old$Eng_little == "little", ]
toplot <- binned
toplot$experience <- toplot$e
toplot$te <- as.integer(as.character(toplot$te))

type <- "exp.cont" #was cont
colnames(toplot)[4]<-"score"
toplot$E <- 1
toplot$E[toplot$Eng_little=="little"]<-E[2]
toplot$E[toplot$Eng_little=="lot"]<-E[1]

if (type=="exp.flat"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.flat(toplot$age[i],toplot$te[i],toplot$E[i],PAR)}}
if (type=="exp.disc"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.disc(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3])}}

lottoplot <- toplot[toplot$Eng_little=="lot",]
littoplot <- toplot[toplot$Eng_little=="little",]
str(lottoplot)
if (type=="exp.cont"){
  for (i in 1:length(lottoplot$score)){
    lottoplot$score[i]<-exp.cont(
      lottoplot$age[i],
      lottoplot$te[i],
      lottoplot$E[i],
      PARlot[1],PARlot[2],PARlot[3],PARlot[4])}}
if (type=="exp.cont"){
  for (i in 1:length(littoplot$score)){
    littoplot$score[i]<-exp.cont(
      littoplot$age[i],
      littoplot$te[i],
      littoplot$E[i],
      PARlit[1],PARlit[2],PARlit[3],PARlit[4])}}

if (type=="exp.disc.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.disc.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3])}}
if (type=="exp.cont.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-exp.cont.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
if (type=="pow.flat"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.flat(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2])}}
if (type=="pow.disc"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.disc(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
if (type=="pow.cont"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.cont(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
if (type=="pow.disc.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.disc.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3])}}
if (type=="pow.cont.x"){for (i in 1:length(toplot$score)){toplot$score[i]<-pow.cont.x(toplot$age[i],toplot$te[i],toplot$E[i],PAR[1],PAR[2],PAR[3],PAR[4])}}
	
# Keep lit and lot as arbitrary names	
toplot <- rbind(littoplot, lottoplot)
toplot$te<-as.factor(toplot$te)
toplot$fac<-as.factor(paste(toplot$Eng_little,toplot$te,sep=""))

type <- "exp.flat"

q<-ggplot(data=toplot[toplot$Eng_little=="lot",],aes(x=age,y=score))+geom_line(aes(color=factor(fac)),alpha=1)+scale_color_manual(values=c("#009E73","#009E73","#009E73","#D55E00","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#D55E00","#D55E00","#D55E00"))
for (i in seq(2,29,3)){
  jitter<-0
  if (type=="exp.flat"){
    numage<-min(toplot$age[toplot$Eng_little=="lot" & toplot$te==i])
  }
  q<-q+annotate("text",x=numage,y=toplot$score[toplot$Eng_little=="lot" & toplot$te==i & (toplot$age==numage)],label=as.character(i),color="black",size=3.5)
}
q<-q+guides(color=FALSE)+labs(y="accuracy (log-odds)",x="age")+theme(plot.title=element_text(size=11,color="black"),axis.title=element_text(size=14,color="black"),axis.text=element_text(size=11,color="black"))
if(logodds){
  q<-q+scale_y_continuous(lim=c(1,4))
}else{
  q<-q+scale_y_continuous(lim=c(.75,1))
}
q<-q+scale_x_continuous(lim=c(0,71)) +  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"))
q

q2<-ggplot(data=toplot[toplot$Eng_little=="little",],aes(x=age,y=score))+geom_line(aes(color=factor(fac)),alpha=1)+scale_color_manual(values=c("#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#009E73","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#56B4E9","#D55E00","#D55E00","#D55E00","#D55E00","#D55E00","#D55E00"))
#for (i in seq(4,30,1)){
#  jitter<-0
#  if (i==5 | i==8){jitter<-11}
#  if (i==9){jitter<-16}
#  if (i==11){jitter<-5}
#  if (i==15){jitter<-4}
#  if (i==22){jitter<-14}
#  if (i==24){jitter<-5}
#  if (i==26){jitter<-11}
#  if (i==29){jitter<-3}
#  	  q2<-q2+annotate("text",x=max(toplot$age[toplot$Eng_little=="little" & toplot$te==i])-jitter,y=toplot$score[toplot$Eng_little=="little" & toplot$te==i & (toplot$age==max(toplot$age[toplot$Eng_little=="little" & toplot$te==i])-jitter)],label=as.character(i),color="black",size=3.5)
#}
q2<-q2+guides(color=FALSE)+labs(y="accuracy (log-odds)",x="age")+theme(plot.title=element_text(size=11,color="black"),axis.title=element_text(size=14,color="black"),axis.text=element_text(size=11,color="black"))
	if(logodds){
  q2<-q2+scale_y_continuous(lim=c(1,4))
}else{
  q2<-q2+scale_y_continuous(lim=c(.75,1))
}
q2
q2<-q2+scale_x_continuous(lim=c(0,71)) +  theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(), panel.background = element_blank(), axis.line = element_line(colour = "black"))
q2
q
	
type <- "exp.cont"	## required for rate plot
if (type=="exp.flat"){rate=data.frame(age=seq(1,70,.1),rate=PAR)}
if (type=="exp.disc" | type=="exp.disc.x"){rate=data.frame(age=seq(1,70,.1),rate=c(rep(PAR[2],round(PAR[1])),rep(PAR[3],70-round(PAR[1]))))}

if (type=="exp.cont"){
	ratelot=data.frame(age=seq(1,70,.1),ratelot=c(rep(PARlot[2],691)))
	ratelot$ratelot[ratelot$age>PARlot[1]]<-sapply(ratelot$age[ratelot$age>PARlot[1]],function(x){PARlot[2]*(1-1/(1+exp(-PARlot[3]*(x-PARlot[1]-PARlot[4]))))})
	ratelit=data.frame(age=seq(1,70,.1),ratelit=c(rep(PARlit[2],691)))
	ratelit$ratelit[ratelit$age>PARlit[1]]<-sapply(ratelit$age[ratelit$age>PARlit[1]],function(x){PARlit[2]*(1-1/(1+exp(-PARlit[3]*(x-PARlit[1]-PARlit[4]))))})
}

if (type=="pow.flat"){rate=data.frame(age=seq(1,70,1),rate=PAR[2])}
if (type=="pow.disc" | type=="pow.disc.x"){rate=data.frame(age=seq(1,70,1),rate=c(rep(PAR[2],round(PAR[1])),rep(PAR[3],70-round(PAR[1]))))}
if (type=="pow.cont" | type=="pow.cont.x"){
	rate=data.frame(age=seq(1,70,1),rate=c(rep(PAR[2],70)))
	rate$rate[rate$age>PAR[1]]<-sapply(rate$age[rate$age>PAR[1]],function(x){PAR[2]*(1-1/(1+exp(-PAR[3]*(x-PAR[1]-PAR[4]))))})
}

#r<-ggplot(data=rate,aes(x=age,y=rate))+geom_line(size=1.5)
#if (type=="exp.disc.x" | type=="exp.cont.x" | type=="pow.disc.x" | type=="pow.cont.x"){
#	r<-r+scale_x_discrete(name="yrs experience",breaks=seq(0,70,10))
#}

r <-ggplot(data=ratelit,aes(x=age,y=ratelit))+geom_line(size=1.5) + scale_color_manual(values=c(rep("light grey",19),"black"))
r <- r + theme(axis.text = element_text(color="black",size=16))
r <- r + xlab("age") + ylab("rate")
r <- r + theme(axis.title.y = element_text(size=15)) + scale_y_continuous(limits = c(0, .20)) # same scale 
r <- r + theme(axis.title.x = element_text(size=15))	
r <- r + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),panel.background = element_blank(), axis.line = element_line(colour = "black"), legend.position="none")
r

r2 <-ggplot(data=ratelot,aes(x=age,y=ratelot))+geom_line(size=1.5) + scale_color_manual(values=c(rep("light grey",19),"black"))
r2 <- r2 + theme(axis.text = element_text(color="black",size=16))
r2 <- r2 + xlab("age") + ylab("rate")
r2 <- r2 + theme(axis.title.y = element_text(size=15)) + scale_y_continuous(limits = c(0, .20)) # same scale
r2 <- r2 + theme(axis.title.x = element_text(size=15))	
r2 <- r2 + theme(panel.grid.major = element_blank(), panel.grid.minor = element_blank(),panel.background = element_blank(), axis.line = element_line(colour = "black"), legend.position="none")
r2

pdf(file=paste("test","all_lot_little",".pdf",sep="_"),w=8,h=11)
vplayout <- function(x,y) viewport(layout.pos.row=x, layout.pos.col=y)
pushViewport(viewport(layout=grid.layout(3,2,heights=c(1,1,.5),widths=c(.5,.5))))
print(p,vp=vplayout(1,1))
print(q,vp=vplayout(2,1))
print(p2,vp=vplayout(1,2))
print(q2,vp=vplayout(2,2))
print(r,vp=vplayout(3,2))
print(r2,vp=vplayout(3,1))
dev.off()

pdf(file=paste("test","main_lot_little",".pdf",sep="_"),w=7,h=7)
vplayout <- function(x,y) viewport(layout.pos.row=x, layout.pos.col=y)
pushViewport(viewport(layout=grid.layout(2,2,heights=c(1,1),widths=c(.5,.5))))
print(p,vp=vplayout(1,1))
print(q,vp=vplayout(2,1))
print(p2,vp=vplayout(1,2))
print(q2,vp=vplayout(2,2))
dev.off()

pdf(file=paste("rate_little",".pdf",sep="_"),w=5.75,h=2.8)
print(r)
dev.off()
pdf(file=paste("rate_lot",".pdf",sep="_"),w=5.75,h=2.8)
print(r2)
dev.off()

pdf(file=paste("rate_lot_little",".pdf",sep="_"),w=7.30,h=5.04)
vplayout <- function(x,y) viewport(layout.pos.row=x, layout.pos.col=y)
pushViewport(viewport(layout=grid.layout(2,2,heights=c(1,1,.5),widths=c(.5,.5))))
print(r,vp=vplayout(2,2))
print(r2,vp=vplayout(2,1))
dev.off()
